add description field to category model and validation; update category routes and controllers for image handling

+++ src/app/modules/category/category.service.ts
@@ -1,14 +1,35 @@
+import httpStatus from "http-status";
+import { QueryBuilder } from "../../builder/QueryBuilder";
+import AppError from "../../errors/AppError";
 import { TCategory } from "./category.interface";
 import { Category } from "./category.model";
 
 const createCategory = async (payload: TCategory) => {
-  const result = (await Category.create(payload)).populate("productId");
+  const result = await Category.create(payload);
   return result;
 };
 
-const getAllCategory = async (payload: Record<string, unknown>) => {
-  const result = await Category.find({ productId: payload.id });
-  return result;
+const getAllCategory = async (query: Record<string, unknown>) => {
+  // Create a new QueryBuilder instance for the product query
+  const productQuery = new QueryBuilder(Category.find(), query)
+    .search(["name"])
+    .sort()
+    .fields()
+    .filter()
+    .paginate();
+
+  // Execute the query to get the paginated results
+  const result = await productQuery.modelQuery;
+
+  // Create a separate query to count the total number of products matching the filter criteria
+  const countQuery = new QueryBuilder(Category.find(), query)
+    .search(["name"])
+    .filter();
+
+  // Execute the count query to get the total count
+  const totalCount = await countQuery.modelQuery.countDocuments();
+
+  return { totalCount, categories: result };
 };
 
 // get a single Category from the database
@@ -22,6 +43,10 @@ const updateACategory = async (
   id: string,
   updateData: Record<string, unknown>
 ) => {
+  const isExist = await Category.findById(id);
+  if (!isExist) {
+    return new AppError(httpStatus.NOT_FOUND, "Category not found!");
+  }
   const result = await Category.findByIdAndUpdate(
     id,
     { $set: updateData },
@@ -31,6 +56,10 @@ const updateACategory = async (
 };
 // delete a Category from the database
 const deleteACategory = async (id: string) => {
+  const isExist = await Category.findById(id);
+  if (!isExist) {
+    return new AppError(httpStatus.NOT_FOUND, "Category not found!");
+  }
   const result = await Category.findByIdAndDelete(id);
   return result;
 };
